# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GaskeunKerja for Business \u2014 MVP",
    "version": "0.1.0"
  },
  "paths": {
...
    "/jobs": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Create Job",
        "operationId": "create_job_jobs_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Jobs",
        "operationId": "list_jobs_jobs_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobListItemOut"
                  },
                  "title": "Response List Jobs Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job",
        "operationId": "get_job_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "jobs"
        ],
        "summary": "Update Job",
        "description": "Round-3 follow-up #9 (2026-07-19): editing an existing job's text fields no longer\nre-extracts/re-stores competencies at all (previously called _extract_and_store_competencies\nhere, same as create_job). Real problem this closes: skill-gap analysis for already-scored\ncandidates is computed against a job's competency list at analysis time \u2014 if HR could freely\nedit that list afterward (this endpoint used to reopen the full add/dismiss/restore review\nmodal on every edit save too, see JobsListPage.tsx), every existing candidate's skill-gap/score\nwould silently go stale relative to the new list, with no auto-invalidation mechanism to catch\nup (a known, previously-documented gap \u2014 see T17-followup's stale-competency-list note).\nCompetencies are now a one-time decision made at CREATE time only (create_job, unchanged) \u2014\nediting a job's description afterward is purely a text update.",
        "operationId": "update_job_jobs__job_id__put",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "jobs"
        ],
        "summary": "Close Job",
        "description": "Soft-delete: sets status='closed'. Never a real SQL DELETE \u2014 avoids FK errors\nagainst candidates/interviews/audit_log and preserves audit history.",
        "operationId": "close_job_jobs__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/permanent": {
      "delete": {
        "tags": [
          "jobs"
        ],
        "summary": "Hard Delete Job",
        "description": "Real, irreversible cascade delete \u2014 separate from the existing soft-close DELETE above\n(kept as-is). Round-2 polish (2026-07-17), user-confirmed tradeoff: this destroys audit\nhistory for the job. No ORM cascade exists anywhere in this schema (confirmed by reading\nevery model file), so the cascade is hand-written here in FK-safe order, plus on-disk\nCV/audio/report file cleanup per candidate.",
        "operationId": "hard_delete_job_jobs__job_id__permanent_delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/reports": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Job Reports",
        "description": "Every GENUINELY invited candidate for this job with their interview/decision pipeline stage\n\u2014 powers the Laporan page (#15 + Round 12/13/14 follow-up). Round 13: candidates who haven't\neven been invited yet are excluded entirely (not just coarsely labeled \"Menunggu Wawancara\") \u2014\nLaporan is about tracking invited candidates through interview -> decision, not surfacing the\nwhole uninvited CV pool, which is the Kandidat page's job.\n\nRound 14 follow-up (real bug, user-reported): `invited_at`/`contact_email` are NOT proof of a\nreal invite \u2014 `invited_at` is set the moment HR opens the invite modal (token generation), and\n`contact_email` can be added/edited long after with no send action ever happening. The only\nreal signal that an invite email was actually dispatched is `invite_email_sent_at`\n(routers/candidates.py::send_candidate_invite_email) \u2014 filtering on that instead mirrors the\nidentical fix in routers/matching.py and CandidateDetailPage.tsx/ShortlistPage.tsx's status\nlogic.",
        "operationId": "list_job_reports_jobs__job_id__reports_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobReportListItemOut"
                  },
                  "title": "Response List Job Reports Jobs  Job Id  Reports Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobCreateRequest": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "responsibilities": {
            "type": "string",
            "title": "Responsibilities"
          },
          "requirements": {
            "type": "string",
            "title": "Requirements"
          },
          "qualifications": {
            "type": "string",
            "title": "Qualifications"
          }
        },
        "type": "object",
        "required": [
          "title",
          "responsibilities",
          "requirements",
          "qualifications"
        ],
        "title": "JobCreateRequest"
      },
...
      "JobOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "company_id": {
            "type": "integer",
            "title": "Company Id"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "responsibilities": {
            "type": "string",
            "title": "Responsibilities"
          },
          "requirements": {
            "type": "string",
            "title": "Requirements"
          },
          "qualifications": {
            "type": "string",
            "title": "Qualifications"
          },
          "required_education_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required Education Level"
          },
          "required_major": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required Major"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_id",
          "company_name",
          "title",
          "responsibilities",
          "requirements",
          "qualifications",
          "required_education_level",
          "required_major",
          "status",
          "created_at"
        ],
        "title": "JobOut"
      },
...
    }
  }
}